requirements.txt
pandas>=1.5.0
python-dotenv>=1.0.0
requests>=2.28.0
boto3>=1.26.0 